Fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,14 +8,15 @@
 
 int main(void)
 {
-	int a = 1;
-	int b = 2;
+	long int a = 1;
+	long int b = 2;
 	int i;
+
 	printf("%d, %d", a, b);
 
 	for (i = 0; i < 48; i++)
 	{
-		int sum = a + b;
+		long int sum = a + b;
 		printf(", %d", sum);
 		a = b;
 		b = sum;
